requirements.txt
# Core dependencies
google-generativeai>=0.3.0
requests>=2.28.0
streamlit>=1.28.0
pandas>=1.5.0
python-dotenv>=0.19.0

# Medical NLP libraries
medspacy>=1.0.0
spacy>=3.4.0
transformers>=4.21.0
torch>=1.12.0
scikit-learn>=1.1.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11

# Data handling and utilities
json5>=0.9.0
pydantic>=1.10.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Development tools
black>=22.0.0
flake8>=5.0.0